Add dir for task 1
