+++ Task1_Phone_Book/Makefile
@@ -1,10 +1,14 @@
 FLAGS = -Werror -Wall -Wextra
 CC = gcc $(FLAGS)
+SOURCE = main.c
 
 all :
-	$(CC) main.c -o phone_book.out
+	$(CC) $(SOURCE) -o phone_book.out
 
 rebuild: clean all
+
+style:
+	clang-format -i $(SOURCE)
 
 clean :
 	rm -fr *.o
